First commit
All the files to date showcasing what I've done so far. I am still learning how to handle @media in order to ensure all elements work on different screen sizes. It turned out to be quite a challenging task to adjust for what I had built, and a useful lesson going forward as a reminder to always begin a project with all screen sizes in mind.